Extract simple classes and types in common.py and create public repository.py interface
