+++ pyproject.toml
@@ -72,3 +72,4 @@ select = ["E", "F", "I"]
 
 [tool.ruff.per-file-ignores]
 "__init__.py" = ["F401"]
+"repository.py" = ["F401"]
