+++ sqlalchemy_bind_manager/__init__.py
@@ -3,12 +3,6 @@ from ._bind_manager import (
     SQLAlchemyBindManager,
     SQLAlchemyConfig,
 )
-from ._repository import (
-    PaginatedResult,
-    SortDirection,
-    SQLAlchemyAsyncRepository,
-    SQLAlchemyRepository,
-)
 from ._unit_of_work import (
     AsyncUnitOfWork,
     UnitOfWork,
