+++ sqlalchemy_bind_manager/_repository/__init__.py
@@ -1,3 +1,4 @@
 from .async_ import SQLAlchemyAsyncRepository
-from .base_repository import PaginatedResult, SortDirection
+from .base_repository import SortDirection
+from .common import Cursor, CursorPaginatedResult, PaginatedResult
 from .sync import SQLAlchemyRepository
