+++ sqlalchemy_bind_manager/_repository/sync.py
@@ -18,14 +18,10 @@ from .._bind_manager import SQLAlchemyBind
 from .._transaction_handler import SessionHandler
 from ..exceptions import InvalidConfig, ModelNotFound
 from .base_repository import (
-    MODEL,
-    PRIMARY_KEY,
     BaseRepository,
-    Cursor,
-    CursorPaginatedResult,
-    PaginatedResult,
     SortDirection,
 )
+from .common import MODEL, PRIMARY_KEY, Cursor, CursorPaginatedResult, PaginatedResult
 
 
 class SQLAlchemyRepository(Generic[MODEL], BaseRepository[MODEL], ABC):
