+++ tests/repository/async_/test_cursor_pagination.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sqlalchemy_bind_manager._repository.base_repository import Cursor
+from sqlalchemy_bind_manager.repository import Cursor
 
 
 def _test_models(model_class):
