+++ tests/unit_of_work/sync/conftest.py
@@ -9,8 +9,8 @@ from sqlalchemy.orm import clear_mappers
 from sqlalchemy_bind_manager import (
     SQLAlchemyBindManager,
     SQLAlchemyConfig,
-    SQLAlchemyRepository,
 )
+from sqlalchemy_bind_manager.repository import SQLAlchemyRepository
 
 
 @pytest.fixture
